requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
aiomysql==0.2.0
alembic==1.13.0
greenlet>=1.1.0
